Add collection size frontend

+++ api/schema.sql
@@ -50,7 +50,7 @@ CREATE TABLE IF NOT EXISTS map (
 
 CREATE TABLE IF NOT EXISTS collections (
     id          BIGSERIAL,
-    size        INTEGER,
+    size        BIGINT,
     name        TEXT UNIQUE NOT NULL,
     sources     JSONB,
     created     TIMESTAMP
